requirements.txt
git+https://github.com/dolfies/discord.py-self
pynacl
python-dotenv
